[이슈#129] 맵과 collectionView 연결
마커와 연결된 collectionView가 보이도록 설정
마커 클릭시 연결되는 카드가 보이도록 설정
