+++ ios_baemin_waiting/ios_baemin_waiting/Helper/Constant.swift
@@ -9,6 +9,6 @@
 import Foundation
 
 let baseOrange = UIColor(red: 255/255, green: 173/255, blue: 39/255, alpha: 1.0)
-//let baseURL = "http://52.78.157.5:8080"
+let baseURL = "http://52.78.157.5:8080"
 
-let baseURL = "http://192.168.100.74:8080/baeminWaiting004/"
+//let baseURL = "http://192.168.100.74:8080/baeminWaiting004"
